Cleaned up server structure
